初步完成收藏与收藏展示功能。

+++ app/src/main/java/com/hnqcgc/redfirecookbook/logic/network/RedFireCookBookNetwork.java
@@ -27,7 +27,7 @@ public class RedFireCookBookNetwork {
         return service.searchAllRecipe(10, offset);
     }
 
-    public Call<RecipeDetails> searchRecipe(int recipeId) {
+    public Call<RecipeDetails> searchRecipe(long recipeId) {
         return service.searchRecipe(recipeId);
     }
